+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Version 0.8 (2020-05-??)
 - feat, l1+l2: split out common code for adding context to boolean column
 - feat, l2: incorporate manually-done datacleaning of confirmed cases + add check that no negatives
 - enh, l2: improvements to the interpolation after testing in notebook
+- feat, l2: drop spikes by easing (from notebook t15b)
+- enh, l0: add brazil
 
 
 Version 0.7 (2020-05-06)